+++ web/src/components/routes.js
@@ -2,11 +2,14 @@ import config from "../app/config";
 import { shortUrl } from "../app/utils";
 
 const routes = {
+  app: config.app_root,
   login: "/login",
   signup: "/signup",
-  app: config.app_root,
   account: "/account",
   settings: "/settings",
+  passwordResetRequest: "/reset-password",
+  passwordReset: "/account/password/reset/:token",
+  emailVerify: "/account/email/verify/:token",
   subscription: "/:topic",
   subscriptionExternal: "/:baseUrl/:topic",
   forSubscription: (subscription) => {